packages/apps/Music: Resume playback only on gaining focus.
- Start Music playback.Start and end voice calls very fast.
Music playback doesn't pause everytime we are starting a voice call.
- When a music playback starts, fadein volume is applied in steps
of 1% in every 10ms to ensure gradual increase in volume level.
play() was called from the messageloop of fadein volume for start
playback usecase.Due to this, when player goes to pause state,
leftover messages for fadein volume was causing music player
to go from pause to playback state.
-Resume playback only after Music app window gains focus.

Change-Id: I79681edb20e7ed3636868e4749dc3738d13e1d2b
CRs-Fixed: 271135