packages/apps/Music: Resume playback only on gaining focus.
- Start Music playback.Start and end voice calls very fast.
Music playback doesn't pause everytime we are starting a voice call.
- When a music playback starts, fadein volume is applied in steps
of 1% in every 10ms to ensure gradual increase in volume level.
play() was called from the messageloop of fadein volume for start
playback usecase.Due to this, when player goes to pause state,
leftover messages for fadein volume was causing music player
to go from pause to playback state.
-Resume playback only after Music app window gains focus.

Change-Id: Ibd668719f62096c9a04607d1b5d5c63871e2cb6d
CRs-Fixed: 261126