Music: Fix to make progress bar update for fast forward and rewind while in suspend
- While playback, press suspend key. While in suspend press fast forward
  continuously. UI screen activates and starts fast forward from the
  beginning of the playback.
- While in suspend, the first fast forward key press is activating the UI
  screen. On second fast forward key press, reset the seek position and
  moves back to starting of the playback.
- Changes are done to seek from the same position of playback while in suspend.

Music: Fix for seekbar issue on interrupt by call

- Seek bar stucks on call interrupt while fast-forwarding
or rewinding a clip.
- When user presses next/previous for seeking, scanForward()
is called by fwdListener which in turn performs seek operation
updating timeline and seekbar. Each event of seek is posted to
the listener with a delay of 260ms.
In this usecase when the target gets the call, by the time seek event
is posted with repcnt as -1, mService is set to NULL.Since mService
is null it returns without setting mPosoverride to -1.Due to this,
mPosoverride retains the older value where the last seek operation
was successfully completed.When call is disconnected and music
playback is resumed, since mPosoverride is positive, position
of the seekbar is always updated with mPosoverride in refreshNow().
- In scanForward() and scanBackward(), set mPosoverride to -1
in case mService is null and is called with repcnt -1.
Since seek event is posted with a delay, this ensures that
last seek event is acknowledged even if mService is null.

packages/apps/Music: Fix for excessive JNI references issue#2

- During monkey run on music app, some JNI references from music
  app are not being dereferenced and resulting framework reboot.
- In the TrackBrowserActivity, the cursor JNI references made
  while creating NowPlayingCursor object are not being closed.
- This fix ensures all the cursor references are closed properly
  to avoid excessive JNI references.

Music: Fix for seek bar stucks after pressing suspend key

- Seek bar stucks while pressing Suspend key during Fast-forwarding
or Rewinding the clip
- During power suspend, mPosoverride value is set to the current
playback position since seeking is in progress. mPosoverride is
reset only when user stops pressing seek button. In this use case
Suspend is called before getting the last seek event. Hence seek bar
is always at a constant position taking the value of mPosoverride.
- Reset mPosoverride to -1 when suspend key is pressed if
mPosoverride is set to a positive value.

Change-Id: Ied2fbc8df67b85050345c78291ba844db09626e6
(cherry picked from commit cb6d9aa1533e1ce097b86a4f9c316f85214cb2cb)