Music: error found when music player play an unsupported file
In the latest music app code, As part of QRD enhancements, a
broadcast receiver has been added for action
ACTION_AUDIO_BECOMING_NOISY in AudioPreview page. This was
registered at the end of oncreate method and unregistered in
ondestroy method.

Now, unsupported music files are handled at beginning of onCreate()
method and music app finishes the activity for such unsupported files.
Thus, apparently broadcast receiver will not get registered and due to
finish() call, onDestroy() will be called which try to unregister the
receiver (which is not registered at all). Resulting in a music app crash

CR-fixed: 513148

Change-Id: I9faead22bf7037b25a5e1cb78842d45e2915489d