Music: Fix music application crash when running monkey test
It's the RuntimeException which causes this issue. The root
cause is too many handlers are running at the same time.
There are many instances of ScanningProgress are created and
the handler will get into infinite loops while cursor is null.

Add counter to limit the retry times to avoid RuntimeException
for fixing this issue.

Change-Id: I705a4b8403d31d673cf1a3c299a6d963a4aa4f14
CRs-Fixed: 586019