Music: Music widget works and displays normal when mount SD card
After SD card mounted, cursor for mPlayPos will be null and
should be return in this case when call play next. But now it
still try to open which causes the toast prompt.
After SD card mounted, directly return if the cursor is null.
Also add the adjustment to check the count of cursor in case of
CursorIndexOutOfBoundsException.

CRs-Fixed: 520874
Change-Id: Ic910cc931759fa3dc72eb8ac6efab0d1faa02044