Music: Fix seekbar does not update when interrupting by call
In this case when the target gets the call, the seek event is posted
with repcnt as -1, mService as NULL. Since mService is null it returns
without setting mPosoverride to -1. Due to this, mPosoverride retains
the older value. Set mPosoverride to -1 and add it both in method
scanForward and scanBackward.

Change-Id: I4097176ed50c14c2d3c0c24f884ab3d143ce2a8c
CRs-Fixed: 602349