SnapdragonMusic: recording name is not right
Recording name is not right in the music after
rename it.

The title columns in MediaStore are always populated
based on what MediaMetadataRetriever returns.
MediaProvider ignore any attempt to override these
values from apps.
So in the Music app, we will use the _diaplay_name
column to replace the title column to show the
correct file name.

Change-Id: Ia6a35d76ffd0ead9961bd96aad874259613bab4f
CRs-Fixed: 2655750