SnapdragonMusic: fix toast shown after locale changed
RootCause: Since Android Q, playing music in call is not allowed.
Thus, one fix has been made to show toast when trying to interact
with the app in call. However, AOSP will update widget and notification
when changing locale since music app is listen to locale change broadcast
to update configuration of widget.As a result, the toast will be shown
even the user has not tried to play music.

Solution: add one input param to control if showing toast is required.
Change-Id: I6c71262ef874337db415ef1751577d58709caaa3